+++ backend/game/lobby/lobby.go
@@ -1,4 +1,4 @@
-package game
+package lobby
 
 import (
 	"errors"
@@ -9,11 +9,12 @@ import (
 )
 
 type Lobby struct {
-	id        string
-	lock      sync.Mutex
-	masterKey string
-	players   map[string]*player
-	game      *logic.Game
+	id            string
+	lock          sync.Mutex
+	masterKey     string
+	players       map[string]*player
+	game          *logic.Game
+	inactiveReset chan struct{}
 }
 
 func (lobby *Lobby) ID() string {
@@ -32,9 +33,12 @@ func (lobby *Lobby) StartGame(key string) error {
 	for _, player := range lobby.players {
 		names = append(names, player.name)
 	}
-	game, err := logic.NewGame(names)
+	game, err := logic.NewGame(names, lobby.inactiveReset)
 	if err != nil {
 		return err
+	}
+	if lobby.inactiveReset != nil {
+		lobby.inactiveReset <- struct{}{}
 	}
 	lobby.game = game
 	return nil
@@ -56,6 +60,10 @@ func (lobby *Lobby) IsLobbyMaster(key string) bool {
 	lobby.lock.Lock()
 	defer lobby.lock.Unlock()
 	return lobby.masterKey == key
+}
+
+func (lobby *Lobby) close() {
+	close(lobby.inactiveReset)
 }
 
 func randomString(n int) string {