+++ backend/game/logic/card.go
@@ -42,13 +42,12 @@ func standardCardSet() []*card {
 	return cards
 }
 
-func (c *card) calculateValue(trickColor, deckColor string) int {
+func (c *card) calculateValue(winningColor string) int {
 	if c.color != "" {
-		if c.color == trickColor {
+		if c.color == winningColor {
 			return c.value + 13
-		} else if c.color != deckColor {
-			return 0
 		}
+		return 0
 	}
 	return c.value
 }
@@ -60,13 +59,6 @@ func getDeckColor(deck []*card) string {
 		}
 	}
 	return ""
-}
-
-func getTrickColor(trickColor string, deck []*card) string {
-	if trickColor == "" {
-		return getDeckColor(deck)
-	}
-	return trickColor
 }
 
 func (c *card) isPlayable(deckColor string, hand map[int]*card) bool {