+++ frontend/src/App.tsx
@@ -1,20 +1,33 @@
 import React from "react"
-import { AppBar, createStyles, Theme, Toolbar, Typography, withStyles, WithStyles } from "@material-ui/core"
+import { AppBar, createStyles, IconButton, Theme, Toolbar, Typography, withStyles, WithStyles } from "@material-ui/core"
 import { Redirect, Route, Switch } from "react-router-dom"
 import { Lobby } from "./lobby"
 import Board from "./board/Board"
 import TrickCallDialog from "./board/TrickCallDialog"
 import Deck from "./board/Deck"
 import Hand from "./board/Hand"
+import { SettingsDialog } from "./settings"
+import { Settings as SettingsIcon } from "@material-ui/icons"
 
 function App(props: WithStyles<typeof styles>) {
+	const [openSettings, setOpenSettings] = React.useState(false)
 	return (
 		<div className={props.classes.root}>
 			<AppBar position="static">
 				<Toolbar>
-					<Typography variant="h6" color="textPrimary">Wizard Online</Typography>
+					<Typography
+						className={props.classes.title}
+						variant="h6"
+						color="inherit">
+						Wizard Online
+					</Typography>
+					<IconButton color="inherit" onClick={() => setOpenSettings(true)}>
+						<SettingsIcon />
+					</IconButton>
 				</Toolbar>
 			</AppBar>
+
+			<SettingsDialog open={openSettings} onClose={() => setOpenSettings(false)} />
 
 			<div className={props.classes.content}>
 				<Switch>
@@ -22,8 +35,7 @@ function App(props: WithStyles<typeof styles>) {
 					<Route path="/lobby"><Lobby /></Route>
 					<Route exact path="/game"><Board /></Route>
 					<Route exact path="/test">
-						<Deck 
-						/>
+						<SettingsDialog open={true} onClose={() => console.log("hi")} />
 					</Route>
 					<Route path="/"><Redirect to="/lobby" /></Route>
 				</Switch>
@@ -43,6 +55,9 @@ const styles = (theme: Theme) => createStyles({
 		backgroundColor: theme.palette.background.default,
 		flexGrow: 1,
 		padding: theme.spacing()
+	},
+	title: {
+		flexGrow: 1
 	}
 })