+++ frontend/src/settings/type.ts
@@ -1,0 +1,9 @@
+export default interface Settings {
+	notifications: {
+		enabled: boolean
+		audio: boolean
+		audioTrack: string
+		desktop: boolean
+		playerTurn: boolean
+	}
+}